+++ crates/multihash/src/atoms.rs
@@ -1,0 +1,5 @@
+rustler::atoms! {
+    ok,
+    error,
+    invalid_code,
+}
