+++ crates/multihash/src/lib.rs
@@ -1,36 +1,20 @@
-use multihash::{Code, MultihashRef};
-use rustler::{Binary, Encoder, Env, NifResult, OwnedBinary, Term};
-use std::{convert::TryFrom, io::Write};
+extern crate rustler;
 
-rustler::atoms! {
-    ok,
-    error,
-    invalid_code,
-}
+use rustler::{Env, Term};
 
-#[rustler::nif]
-fn digest<'a>(env: Env<'a>, data: Binary, raw_code: u64) -> NifResult<Term<'a>> {
-    match Code::try_from(raw_code) {
-        Ok(digest_code) => {
-            let digest = digest_code.digest(data.as_slice());
-            let mut binary = OwnedBinary::new(digest.len()).unwrap();
-            binary.as_mut_slice().write_all(&digest).unwrap();
-            Ok((ok(), Binary::from_owned(binary, env)).encode(env))
-        }
-        Err(_) => Ok((error(), invalid_code()).encode(env)),
-    }
-}
+mod atoms;
+mod multihash;
 
-#[rustler::nif]
-fn code<'a>(env: Env<'a>, data: Binary) -> NifResult<Term<'a>> {
-    match MultihashRef::from_slice(data.as_slice()) {
-        Ok(hash) => Ok((ok(), u64::from(hash.algorithm())).encode(env)),
-        Err(_) => Ok((error(), invalid_code()).encode(env)),
-    }
-}
-
-pub fn on_load(_env: Env, _load_info: Term) -> bool {
+fn load(env: Env, _: Term) -> bool {
+    multihash::load(env);
     true
 }
 
-rustler::init!("multihash_nif", [digest, code], load = on_load);
+rustler::init!(
+    "erlang_multihash",
+    [
+        multihash::code,
+        multihash::digest,
+    ],
+    load=load
+);
