+++ crates/multihash/src/multihash.rs
@@ -1,0 +1,29 @@
+use multihash::{Code, MultihashRef};
+use rustler::{Binary, Encoder, Env, NifResult, OwnedBinary, Term};
+use std::{convert::TryFrom, io::Write};
+use crate::atoms::{ok, error, invalid_code};
+
+#[rustler::nif]
+fn digest<'a>(env: Env<'a>, data: Binary, raw_code: u64) -> NifResult<Term<'a>> {
+    match Code::try_from(raw_code) {
+        Ok(digest_code) => {
+            let digest = digest_code.digest(data.as_slice());
+            let mut binary = OwnedBinary::new(digest.len()).unwrap();
+            binary.as_mut_slice().write_all(&digest).unwrap();
+            Ok((ok(), Binary::from_owned(binary, env)).encode(env))
+        }
+        Err(_) => Ok((error(), invalid_code()).encode(env)),
+    }
+}
+
+#[rustler::nif]
+fn code<'a>(env: Env<'a>, data: Binary) -> NifResult<Term<'a>> {
+    match MultihashRef::from_slice(data.as_slice()) {
+        Ok(hash) => Ok((ok(), u64::from(hash.algorithm())).encode(env)),
+        Err(_) => Ok((error(), invalid_code()).encode(env)),
+    }
+}
+
+pub fn load(_env: Env) -> bool {
+    true
+}
